Remove constant bulk viscosity option and fix zero-viscosity checks.
The model runs ideal evolution equations if the viscosities are zero.
However, this was checked many times throughout the code and no longer made
sense with temperature-dependent viscosities.  Now, it is checked once during
initilization and saved in boolean variables so the logic is consistent through
the full time evolution.